ENH: Handle multiple IMPLICIT_DEPENDS files, improve docs and testing
The code handling IMPLICIT_DEPENDS was only able to track a single file,
the latest file replaced earlier files in the list.

The documentation now mentions that the language has to be prefixed to
every file and the test now uses two implicit dependencies, where only
the second is modified to trigger re-running of the custom command.

Fixes #10048.

Signed-off-by: Michael Wild <themiwi@users.sourceforge.net>